// api/src/products/enums/product.enums.ts
export enum ProductStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  DISCONTINUED = 'DISCONTINUED',
}
export enum UnitOfMeasure {
  UN = 'UN',
  KG = 'KG',
  G = 'G',
  L = 'L',
  ML = 'ML',
  M = 'M',
  CM = 'CM',
  M2 = 'M2',
  M3 = 'M3',
  CX = 'CX',
  PCT = 'PCT',
  PAR = 'PAR',
}
export enum MerchandiseOrigin {
  NATIONAL = '0',
  DIRECT_IMPORT = '1',
  INTERNAL_FOREIGN = '2',
  NATIONAL_IMPORT_40_70 = '3',
  NATIONAL_BASIC_PROCESS = '4',
  NATIONAL_IMPORT_UNDER_40 = '5',
  DIRECT_IMPORT_NO_SIMILAR = '6',
  INTERNAL_FOREIGN_NO_SIMILAR = '7',
  NATIONAL_IMPORT_OVER_70 = '8',
}